\name{fertilmodel-package}
\alias{fertilmodel-package}
\docType{package}
\title{
This is an R package that fits 4 fertility models.
}
\description{
Four fertility models are fitted using non-linear least squares. These are the Hadwiger, the Gamma, the Model1 and Model2, following the terminology of the following paper: Peristera P. and Kostaki A. (2007). Modeling fertility in modern populations. Demographic Research, 16(6): 141--194. <doi:10.4054/DemRes.2007.16.6>.
}
\details{
\tabular{lll}{
Package: \tab nlgmcr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2024-11-29\cr
License: \tab GPL-2\cr
}
}
\author{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

\section{Maintainers }{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

\references{
Peristera P. and Kostaki A. (2007). Modeling fertility in modern populations. Demographic Research, 16(6), 141--194. <doi:10.4054/DemRes.2007.16.6>.
}

\note{
\bold{Acknowledgments:}
This package is dedicated to Sanaa who introduced me to these models and whom I may never see again.
}

\keyword{ fertility models }



